feat: add auth

+++ src/users/users.controller.ts
@@ -9,6 +9,9 @@ import {
   ParseUUIDPipe,
   UseInterceptors,
   ClassSerializerInterceptor,
+  NotFoundException,
+  HttpCode,
+  HttpStatus,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -36,7 +39,12 @@ export class UsersController {
   @UseInterceptors(ClassSerializerInterceptor)
   @Get(':id')
   findOne(@Param('id', ParseUUIDPipe) id: string): ResponseUserDto {
-    const user = this.usersService.findOne(id);
+    const user = this.usersService.findById(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     return new ResponseUserDto(user);
   }
 
@@ -47,11 +55,19 @@ export class UsersController {
     @Body() passwordData: UpdatePasswordDto,
   ): ResponseUserDto {
     const user = this.usersService.update(id, passwordData);
+
     return new ResponseUserDto(user);
   }
 
   @Delete(':id')
+  @HttpCode(HttpStatus.NO_CONTENT)
   remove(@Param('id', ParseUUIDPipe) id: string) {
+    const user = this.usersService.findById(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     return this.usersService.remove(id);
   }
 }
